backend/requirements.txt
fastapi[all]==0.100.1
uvicorn[standard]==0.23.2
python-dotenv==1.0.0

starlette==0.27.0
Jinja2==3.1.2
Pillow==9.5.0
openai==0.27.6
ipython==8.14.0
together==0.2.6